rename config dump to list, remove original config list
