+++ src/index.js
@@ -12,6 +12,5 @@ export { default as generateWebpack } from './generate-webpack'
 export { default as configSet } from './config-set'
 export { default as configRemove } from './config-remove'
 export { default as configList } from './config-list'
-export { default as configDump } from './config-dump'
 export { default as configSync } from './config-sync'
 export { version } from '../package.json'
